requirements.txt
Flask==0.10.1
PyYAML==3.11
requests>=2.7.0
psutil
